detect whether there is an upgrade to run or not

+++ files/updatescript/mediawiki_dbupdate.rb
@@ -8,21 +8,30 @@ def update_php(dir)
   old_dir = Dir.getwd
   Dir.chdir(dir)
   stat = File.stat(dir)
+  ran = false
   sudo(stat.uid, stat.gid) do
-    run('git pull -q && git submodule -q sync && git submodule -q update --init')
-    # make sure we do not have a tampered update_command before running it
-    if File.exists?(update_cmd_file = File.expand_path("#{dir}/../data/php_update_command")) && ((s=File.stat(update_cmd_file)).uid == 0) && (sprintf("%o",s.mode) == "100644")
-      run("MW_INSTALL_PATH=#{dir} #{File.read(update_cmd_file)}")
+    if (`git fetch -q && git log --oneline HEAD..@{u} | wc -l`.split("\n").first.to_i != 0) || (ENV['FORCE_UPGRADE'].to_i == 1)
+      run('git pull -q && git submodule -q sync && git submodule -q update --init')
+      # make sure we do not have a tampered update_command before running it
+      if File.exists?(update_cmd_file = File.expand_path("#{dir}/../data/php_update_command")) && ((s=File.stat(update_cmd_file)).uid == 0) && (sprintf("%o",s.mode) == "100644")
+        run("MW_INSTALL_PATH=#{dir} #{File.read(update_cmd_file)}")
+      else
+        run("MW_INSTALL_PATH=#{dir} php #{dir}/maintenance/update.php --quick --conf #{dir}/LocalSettings.php")
+      end
+      ran = true
     else
-      run("MW_INSTALL_PATH=#{dir} php #{dir}/maintenance/update.php --quick --conf #{dir}/LocalSettings.php")
+      puts "No upgrade present and FORCE_UPGRADE not set to 1, skipping..."
     end
   end
-  # history folder is owned by the run user
-  d = File.join(dir,'cache','history')
-  if File.directory?(d)
-    stat = File.stat(d)
-    sudo(stat.uid, stat.gid) do
-      run("find #{File.join(dir,'cache')} -name '*.html' -type f -delete")
+
+  if ran
+    # history folder is owned by the run user
+    d = File.join(dir,'cache','history')
+    if File.directory?(d)
+      stat = File.stat(d)
+      sudo(stat.uid, stat.gid) do
+        run("find #{File.join(dir,'cache')} -name '*.html' -type f -delete")
+      end
     end
   end
   Dir.chdir(old_dir)
